refactor: split out pyromod methods from client.py
