+++ hydrogram/methods/pyromod/stop_listener.py
@@ -1,0 +1,52 @@
+#  Hydrogram - Telegram MTProto API Client Library for Python
+#  Copyright (C) 2020-present Cezar H. <https://github.com/usernein>
+#  Copyright (C) 2023-present Amano LLC <https://amanoteam.com>
+#
+#  This file is part of Hydrogram.
+#
+#  Hydrogram is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU Lesser General Public License as published
+#  by the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#
+#  Hydrogram is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU Lesser General Public License for more details.
+#
+#  You should have received a copy of the GNU Lesser General Public License
+#  along with Hydrogram.  If not, see <http://www.gnu.org/licenses/>.
+
+import inspect
+
+import hydrogram
+from hydrogram.errors import (
+    ListenerStopped,
+)
+from hydrogram.types import Listener
+from hydrogram.utils import PyromodConfig
+
+
+class StopListener:
+    async def stop_listener(self: "hydrogram.Client", listener: Listener):
+        """
+        Stops a listener, calling stopped_handler if applicable or raising ListenerStopped if throw_exceptions is True.
+
+        :param listener: The :class:`pyromod.types.Listener` to stop.
+        :return: ``void``
+        :raises ListenerStopped: If throw_exceptions is True.
+        """
+        self.remove_listener(listener)
+
+        if listener.future.done():
+            return
+
+        if callable(PyromodConfig.stopped_handler):
+            if inspect.iscoroutinefunction(PyromodConfig.stopped_handler.__call__):
+                await PyromodConfig.stopped_handler(None, listener)
+            else:
+                await self.loop.run_in_executor(
+                    None, PyromodConfig.stopped_handler, None, listener
+                )
+        elif PyromodConfig.throw_exceptions:
+            listener.future.set_exception(ListenerStopped())
